+++ src/layers/fp64/line-layer/line-layer-64.js
@@ -17,12 +17,12 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-import {Layer} from '../../../lib';
-import {assembleShaders} from '../../../shader-utils';
-import {GL, Model, Geometry} from 'luma.gl';
-import {fp64ify} from '../../../lib/utils/fp64';
 
-const glslify = require('glslify');
+import {Layer, assembleShaders} from '../../..';
+import {fp64ify} from '../../../lib/utils/fp64';
+import {GL, Model, Geometry} from 'luma.gl';
+import {readFileSync} from 'fs';
+import {join} from 'path';
 
 const DEFAULT_COLOR = [0, 255, 0, 255];
 
@@ -60,8 +60,18 @@ export default class LineLayer64 extends Layer {
 
     const {attributeManager} = this.state;
     attributeManager.addInstanced({
-      instanceSourcePositionsFP64: {size: 4, update: this.calculateInstanceSourcePositions},
-      instanceTargetPositionsFP64: {size: 4, update: this.calculateInstanceTargetPositions},
+      instanceSourcePositionsFP64: {
+        size: 4,
+        update: this.calculateInstanceSourcePositions
+      },
+      instanceTargetPositionsFP64: {
+        size: 4,
+        update: this.calculateInstanceTargetPositions
+      },
+      instanceElevations: {
+        size: 2,
+        update: this.calculateInstanceElevations
+      },
       instanceColors: {
         size: 4,
         type: GL.UNSIGNED_BYTE,
@@ -73,24 +83,30 @@ export default class LineLayer64 extends Layer {
   draw({uniforms}) {
     const {gl} = this.context;
     const lineWidth = this.screenToDevicePixels(this.props.strokeWidth);
-    const oldLineWidth = gl.getParameter(GL.LINE_WIDTH);
     gl.lineWidth(lineWidth);
     this.state.model.render(uniforms);
-    gl.lineWidth(oldLineWidth);
+    // Setting line width back to 1 is here to workaround a Google Chrome bug
+    // gl.clear() and gl.isEnabled() will return GL_INVALID_VALUE even with
+    // correct parameter
+    // This is not happening on Safari and Firefox
+    gl.lineWidth(1.0);
+  }
+
+  getShaders() {
+    return {
+      vs: readFileSync(join(__dirname, './line-layer-vertex.glsl'), 'utf8'),
+      fs: readFileSync(join(__dirname, './line-layer-fragment.glsl'), 'utf8'),
+      fp64: true,
+      project64: true
+    };
   }
 
   createModel(gl) {
     const positions = [0, 0, 0, 1, 1, 1];
-
     return new Model({
       gl,
       id: this.props.id,
-      ...assembleShaders(gl, {
-        vs: glslify('./line-layer-vertex.glsl'),
-        fs: glslify('./line-layer-fragment.glsl'),
-        fp64: true,
-        project64: true
-      }),
+      ...assembleShaders(gl, this.getShaders()),
       geometry: new Geometry({
         drawMode: GL.LINE_STRIP,
         positions: new Float32Array(positions)
@@ -119,6 +135,19 @@ export default class LineLayer64 extends Layer {
       const targetPosition = getTargetPosition(object);
       [value[i + 0], value[i + 1]] = fp64ify(targetPosition[0]);
       [value[i + 2], value[i + 3]] = fp64ify(targetPosition[1]);
+      i += size;
+    }
+  }
+
+  calculateInstanceElevations(attribute) {
+    const {data, getSourcePosition, getTargetPosition} = this.props;
+    const {value, size} = attribute;
+    let i = 0;
+    for (const object of data) {
+      const sourcePosition = getSourcePosition(object);
+      const targetPosition = getTargetPosition(object);
+      value[i + 0] = sourcePosition[2] || 0;
+      value[i + 1] = targetPosition[2] || 0;
       i += size;
     }
   }