+++ CHANGELOG.md
@@ -22,6 +22,58 @@ Ref: http://keepachangelog.com/en/0.3.0/
 
 ### deck.gl v3
 
+#### [3.1.1] - 2016-12-1
+
+- FIX: ExtrudedChoroplethLayer64 - Now updates uniforms when props change
+- HOTFIX ScatterplotLayer: Fix instancePositions regression in 3.1.0
+- HOTFIX DeckGL: Fix PropType warning regression in 3.1.0 (layers/effects props)
+
+#### [3.1.0] - 2016-11-30
+
+- LineLayer: Support elevations (z coords on positions)
+- LineLayer64: Support elevations (z coords on positions)
+- Scatterplot: Enable updateTriggers on instanceRadius, separate from positions.
+- EXPERIMENTAL FEATURE: ReflectionEffect
+- INTERNAL: Target build env moved to Node 6. Travis tests now run on 6 & 7.
+
+#### [3.0.9, 3.0.10] - 2016-11-18
+
+- Bumped viewport-mercator-project version for the unproject fix.
+- Updated dependency versions (#229)
+- Added test code for attribute-manager (#213)
+- Added minMax radius for the scatterplot-layer (#230)
+
+#### [3.0.8] - 2016-11-18
+
+- Performance optimization: remove unnecessary gl.getParameter() calls (#227)
+
+#### [3.0.7] - 2016-11-16
+
+- Fixed precision issue for vec2_mix_fp64 function (#223)
+- Added mouse event object to hover and click event parameter (#255)
+
+#### [3.0.6] - 2016-11-14
+
+- Bug fix for picking not returning x, y coordinates (#220)
+
+#### [3.0.5] - 2016-11-14
+
+- Performance optimization: prevent redundant data calculation (#222)
+
+#### [3.0.3] - 2016-11-14
+
+- Bug fixes for ScatterplotLayer not being updated with new data prop (#215)
+- Performance optimization: prevent fbo from reallocating every time (#217)
+
+#### [3.0.2] - 2016-11-12
+
+- Performance optimization for examples in gh-pages
+
+#### [3.0.1] - 2016-11-11
+
+- Enable blending by default
+- Fix syntax highlighting in gh-pages
+
 #### [3.0.0] - Major deck.gl Release
 
 For details see [What's New](docs/whats-new.md)